File rename and add docstring
